refact: move SerializableList-related test structures to list package
